+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
-## [Unreleased]
+## 0.2.0 
+
+- Implement HoverProvider for symbols in starlark files.  If the word is a
+  builtin function call a tooltip is provided with a link to the bazel
+  documentation.
+
+## 0.1.0
 
 - Initial release